tiny bit of code clean up
